news/requirements.txt
requests==2.21.0
Flask_Cors==3.0.8
Flask==1.0.2
nltk==3.4.5
pymongo==3.9.0
python-dotenv==0.12.0
redis==3.4.1
dnspython